Use encoder to serialize characteristics.
