+++ src/sentry/similarity/encoder.py
@@ -4,8 +4,6 @@ from collections import (Mapping, Set, Sequence)
 
 import six
 
-from sentry.interfaces.stacktrace import Frame
-
 
 class Encoder(object):
     def __init__(self, types):
@@ -16,7 +14,6 @@ class Encoder(object):
             if isinstance(value, cls):
                 value = function(value)
 
-        # TODO: This should also handle numbers.
         if isinstance(value, six.binary_type):
             return value
         elif isinstance(value, six.text_type):
@@ -50,29 +47,3 @@ class Encoder(object):
             )
         else:
             raise TypeError('Unsupported type: {}'.format(type(value)))
-
-
-def get_frame_attributes(frame):
-    attributes = {}
-
-    if frame.function in set(['<lambda>', None]):
-        attributes['signature'] = (
-            (frame.pre_context or [])[-5:] +
-            [frame.context_line] +
-            (frame.post_context or [])[:5]
-        )
-    else:
-        attributes['function'] = frame.function
-
-    for name in ('module', 'filename'):
-        value = getattr(frame, name)
-        if value:
-            attributes[name] = value
-            break
-
-    return attributes
-
-
-encoder = Encoder({
-    Frame: get_frame_attributes,
-})
